Differentiate between type and property (value) errors in cmap
This is needed so we can generate error OCPP RPCERROR messages with the
correct(ish) error reason. I say "ish" because there is some ambiguity
surrounding lists and whether an incorrectly typed list item should result in a
type error or a value error. The quality of the error reasons is also limited by
Erlang's equivalent treatment of strings and lists of integers. Regardless of
the limits, these errors should be "good enough," at least to start.