Some small property-based tests for ocpp messages
Very basic tests of the encode/decode functionality as well as generators for
four message types. These tests may be replicated for all message types, but I'm
only planning to add them as needed from this point. Eventually there should be
a full set in order to provide coverage of the generated code in ocpp_message_*;
however, writing these is tedious and I want to move on to more interesting
functionality. As more message types are used we will continue to add coverage.