Implement station provisioning from connected through boot pending
This implements handling of BootNotificationRequest messages in the 'connected'
and 'provisioning' states. The happy path is implemented and tested for the
sequence

station -> "BootNotificationRequest"  -> CSMS
station <- "BootNotificationResponse" <- CSMS (status = Pending)

In addition to that path, there are two tests for out-of-order/disallowed
messages sent to the CSMS prior to the BootNotificationResponse being sent.

I'll probably split those tests up eventually.

Some substantial infrastructure is implemented in the test_station module to
enable testing sequences of operations as separate independent tests. I don't
know that this is a good idea, but I like being able to separate each operation
into as separate test while the connection process stays alive.