First cut of station provisioning
This implements most of the requirements in section B of the ocpp 2.0.1
standard. It still needs more tests, the most significant of which are noted in
the comments of test_station.erl.

I have abandoned the separate ocpp_station states for when the CSMS has made a
call but not yet received a response from the station. These are a cute idea,
but there is too much "normal" behavior that needs to be supported while in
these states (e.g. responding to messages from the station). Because they just
seemed to add complexity I am deleting them and will have to prevent multiple
sends by looking at the state machine data.