Remove TODO for test that is not valid
I still think the standard isn't clear, but the fact that the standard lays out
a requirement for how a station MUST respond to certain messages not on this
list leads me to conclude that all messages should be allowed. It won't be hard
to restrict the messages in the future if I am wrong about that.