CSMS SHALL NOT send any messages after station has been rejected
As implemented this applies to all cases when the station has not been either
accepted or put into the pending state. This may not be precisely per the
standard since no requirement appears to state that the CSMS shall not send
messages except after rejection. There is, however, no apparent "MAY" or "SHALL"
send requirements and requiring a boot to at least the Pending state makes
implementation much simpler. I can't imagine this will preclude any normal
operation of a CSMS or charging station.